数据导出导入支持统计与修复

+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/data/work/ExportDataWork.kt
@@ -17,10 +17,10 @@ import indi.dmzz_yyhyy.lightnovelreader.data.json.AppUserDataJsonBuilder
 import indi.dmzz_yyhyy.lightnovelreader.data.json.toJsonData
 import indi.dmzz_yyhyy.lightnovelreader.data.local.room.dao.UserDataDao
 import indi.dmzz_yyhyy.lightnovelreader.data.local.room.entity.UserDataEntity
+import indi.dmzz_yyhyy.lightnovelreader.data.statistics.StatsRepository
 import indi.dmzz_yyhyy.lightnovelreader.data.userdata.UserDataPath
 import indi.dmzz_yyhyy.lightnovelreader.data.web.WebBookDataSource
 import java.io.File
-import java.io.FileNotFoundException
 import java.io.FileOutputStream
 import java.io.IOException
 import java.util.zip.ZipEntry
@@ -33,16 +33,31 @@ class ExportDataWork @AssistedInject constructor(
     private val webBookDataSource: WebBookDataSource,
     private val bookshelfRepository: BookshelfRepository,
     private val bookRepository: BookRepository,
+    private val statsRepository: StatsRepository,
     private val userDataDao: UserDataDao
 ) : Worker(appContext, workerParams) {
-    @Suppress("DuplicatedCode")
+
     override fun doWork(): Result {
         val fileUri = inputData.getString("uri")?.let(Uri::parse) ?: return Result.failure()
         val exportBookshelf = inputData.getBoolean("exportBookshelf", false)
         val exportReadingData = inputData.getBoolean("exportReadingData",false)
         val exportSetting = inputData.getBoolean("exportSetting", false)
-        val exportBookmark = inputData.getBoolean("exportBookmark", false)
-        val json = AppUserDataJsonBuilder()
+
+        return try {
+            val json = buildJson(exportBookshelf, exportReadingData, exportSetting)
+            createFile(fileUri, json)
+        } catch (e: Exception) {
+            e.printStackTrace()
+            Result.failure()
+        }
+    }
+
+    private fun buildJson(
+        exportBookshelf: Boolean,
+        exportReadingData: Boolean,
+        exportSetting: Boolean
+    ): String {
+        return AppUserDataJsonBuilder()
             .data {
                 webDataSourceId(webBookDataSource.id)
                 if (exportBookshelf) {
@@ -61,6 +76,8 @@ class ExportDataWork @AssistedInject constructor(
                     userDataDao.getEntity(UserDataPath.ReadingBooks.path)
                         ?.toJsonData()
                         ?.let(::userData)
+                    statsRepository.getAllReadingStats()
+                        .forEach(::dailyReadingData)
                 }
                 if (exportSetting) {
                     userDataDao.getGroupValues(UserDataPath.Reader.path)
@@ -76,28 +93,24 @@ class ExportDataWork @AssistedInject constructor(
             }
             .build()
             .toJson()
-        return createFile(fileUri, json)
     }
 
     private fun createFile(fileUri: Uri, json: String): Result {
         return try {
-            File(applicationContext.filesDir, "data").let { if (!it.exists()) it.mkdir() }
+            File(applicationContext.filesDir, "data").apply { if (!exists()) mkdir() }
             if (fileUri.scheme.equals("file")) {
                 val file = fileUri.toFile()
-                if (file.exists())
-                    file.delete()
+                if (file.exists()) file.delete()
                 file.createNewFile()
             }
-            applicationContext.contentResolver.openFileDescriptor(fileUri, "w")?.use { parcelFileDescriptor ->
-                ZipOutputStream(FileOutputStream(parcelFileDescriptor.fileDescriptor)).use {
+            applicationContext.contentResolver.openFileDescriptor(fileUri, "w")?.use { descriptor ->
+                ZipOutputStream(FileOutputStream(descriptor.fileDescriptor)).use {
                     it.putNextEntry(ZipEntry("data.json"))
                     it.write(json.toByteArray())
+                    it.closeEntry()
                 }
             }
             Result.success()
-        } catch (e: FileNotFoundException) {
-            e.printStackTrace()
-            Result.failure()
         } catch (e: IOException) {
             e.printStackTrace()
             Result.failure()
